improvement: add deletion script
Add deletion script so that the stack can be easily cleaned up when not
needed anymore. Further, the script adjusts the region automatically, so only
one script is needed for Ireland and Virginia. Also, a name tag is added to
the resources created by the script. Moreover, useful commands and resource
ids are saved into a text file.

Also relates to https://github.com/fastai/courses/issues/6.