aws-alias.sh:  Add aws-nb alias to open the notebook from the command line.
By adding an aws-nb command, it makes it even less necessary to have a
static IP (elastic IP) address for the servers.

Mac, Linux and cygwin all have different commands to do the
'desktop-open' thing, so this figures out which OS you're on, and uses
the right command to open the default web browser to the right IP
address.