move model.flatten above so that model compiles
Currently, the model wouldn't compile for sizes other than (224,224) when fine tuning as under the branch if not include_top, the model is not yet flattened but later in the create function, a dense layer is added (without flattening), hence, raising an error. Changing the position of the model.flatten fixes this.